+++ packages/ui/.svelte-kit/__package__/components/badge/Badge.svelte.d.ts
@@ -1,0 +1,5 @@
+import "./badge.css";
+import type { BadgeProps } from "./props.ts";
+declare const Badge: import("svelte").Component<BadgeProps, {}, "">;
+type Badge = ReturnType<typeof Badge>;
+export default Badge;